Move README and LICENSE to .github
The simple R Markdown static site generator builds any Markdown file
into a webpage. It was adding a README.html file, and another
LICENSE.txt file, in the docs/ directory. Adding these files to the
excludes in _site.yml prevents copying of the LICENSE.txt, but not
README.md. R Markdown seems to build any Markdown file in the root dir.
To avoid this, I moved the LICENSE and README to the .github/ directory
where they will still be detected by GitHub, but not by R Markdown.