Use factors for glht and PMCMRplus variables
I was seeing this inscrutable error with glht:

```text
Error in mcp2matrix(model, linfct = linfct):
Variable(s) 'diet' of class 'character' is/are not contained as a factor
in 'model'.
```

The answer didn't come from the docs, but from Stack Overflow:
https://stackoverflow.com/questions/62812953

I saw a similar error with the Kruskal-Wallis test in PMCMRplus. Answer:
https://stackoverflow.com/questions/18139195